requirements.txt
# Flask 核心依赖
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6

# WebSocket 支持
eventlet==0.33.3
python-socketio==5.9.0

# 图片处理
Pillow==10.0.1

# 中文处理
pypinyin==0.49.0

# 系统监控
psutil==5.9.6

# 开发工具
python-dotenv==1.0.0

# 日志
colorlog==6.7.0

# 数据验证
marshmallow==3.20.1

# 时间处理
python-dateutil==2.8.2

# 文件监控
watchdog==3.0.0

# 缓存
cachetools==5.3.2

# HTTP请求
requests==2.31.0

# JSON处理
ujson==5.8.0

# 配置管理
pydantic==2.4.2
pydantic-settings==2.0.3

# 性能监控
memory-profiler==0.61.0

# 测试工具（开发时使用）
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0